cmd/wavelet: fix unspecified wallet flag not generating random wallet
The documentation states that by not specifying the --wallet flag, a new
wallet will be randomly generated[1]. But currently,  not specifying
the --wallet flag will make Wavelet use config/wavelet.txt as the
wallet. In order to to generate a random wallet, the --wallet flag has
to be explicitly set to empty.

[1] https://wavelet.perlin.net/docs/setup#wallet-management