utils: fix failing TestSelectPeers (#141)
* utils: fix failing TestSelectPeers

The test is suspected to fail intermittently because sometimes when
some of the nodes are closed, grpc.Client doesn't have time to update
its internal state to reflect the closed connection. This affects the
outcome of the test as it depends on client.GetState().

* utils, store: use inmem kv for tests

Using inmem speeds up the tests, and also prevents the tests from
failing on Windows due to file locks.